ADD: BookRepository without Abstract class as Domain Interface

+++ books/repositories/book_repository.py
@@ -1,33 +1,30 @@
 from books.models.book_model import Book
-from books.serializers.book_serializers import BookSerializer
 
 
 class BookRepository:
-    def find_all(self, order_by='id'):
+    def find_all(self, order_by="id"):
         queryset = Book.objects.all()
         if order_by:
             queryset = queryset.order_by(order_by)
         return queryset
 
-    def find_one(self, book_id):
-        return Book.objects.get(pk=book_id)
+    def find_one(self, book_id) -> Book | None:
+        # return Book.objects.get(pk=book_id) # throws exception
+        return Book.objects.filter(pk=book_id).first()
 
-    def create(self, data):
-        serializer = BookSerializer(data=data)
-        if serializer.is_valid():
-            serializer.save()
-            return serializer.data
-        return serializer.errors
+    def create(self, data) -> Book:
+        return Book.objects.create(**data)
 
-    def update(self, book_id, data):
+    def update(self, book_id, data) -> Book:
         book = Book.objects.get(pk=book_id)
-        serializer = BookSerializer(book, data=data)
-        if serializer.is_valid():
-            serializer.save()
-            return serializer.data
-        return serializer.errors
+        for key, value in data.items():
+            setattr(book, key, value)
+        book.save()
+        return book
 
-    def delete(self, book_id):
-        book = Book.objects.get(pk=book_id)
+    def delete(self, book_id) -> bool:
+        book = Book.objects.filter(pk=book_id).first()
+        if not book:
+            return False
         book.delete()
         return True
